RPS-145 [PDP] "Success add" message appears even if you add 0 qty - If a user enters '0' for quantity, the 'Add To Cart' ( A2C ) button is now disabled.   I also discovered, a user could have entered a non-numeral, a letter or character and the item would seem to be added to the cart - but it never would be - still, bad practice so I put a check in to prevent that. So now, only numbers are allowed. Anything else would disable the A2C.
And it a product had variants ( Original Floating ), it was possible to
enter '0' for quantity, see the A2C disabled but switching to a new
variant, the quantity would remain '0' but the A2C was active. Now, if
the quantity is '0' and one switches variants, the A2C is disabled. 

I've used the 'keyup' function so it will work on mobile ( touchscreen )
devices