RPS-145  [PDP] "Success add" message appears even if you add 0 qty There is a separate function where, if the A2C button is clicked, the hop-up showing the item was added to the cart would still display even though nothing is added to the cart. Not sure why this function is not part of the A2C process. None-the-less, I put a condition on it where, if the value is greater than 0, then display the hop-up.
Looking at the video, I did see another flaw. If a user entered in a
value ending in 0, such as 10, the code was only looking for the last
key entry.  I did that to block out non-numerical values. I added to it
a check to see if the quantity selected is greater than 0. So now, a
user can select a quantity of 10