RPS-365
The commit, "3dbcbae", gave ".pt_product-search-result #main" a property
of "z-index:1". Because of this, any time the mobile main menu was used
( when the browser was 959px of less in width ), the "headerwrapper"
would "disappear" because it was now behind the
".pt_product-search-result #main". I've given  the "headerwrapper" an
"z-index" of "2" and this fixes the issue. 

This issue is unrelated to the original ticket.

Also fixed the issue with the breadcrumbs on the PDP. The float was not
giving the breadcrumbs proper height so the box-shadow was only
appearing in the reduced height.