RPS-361
Went through the list of required translations. Fixed them all. One is a
change to the content slot, "shipping-msg". 

As for changing the label for credit card on the billing page, this is
set in BM. I am not making the change here but till put it in the ticket
as not to affect how the credit cards are set up on other instances.