RPS-430 EU - UK Locale for EU site
Added the UK to the EU site. 

Added the UK to the locales. 

I also created a new organizational preference, "ActiveCountries". This
holds a JSON object which lists all the active sites. I created this
because of the Interstital window. This way, the client can rearrange
the order of the flags and there is now only one preference to change
when turning on/off a site. Also, the interstitial window, the layout is
two flags across unless there is one and that one must be centered. I
couldn't think of a better way to make this responsive to this layout
than with the JSON object. 

Also, when a user switches from the EU ( or Finland or any future EU
country ) to the UK, the currency will switch from Euro to the Pound -
and back. 

I did need to add the UK to the site Alias file in the same was as was
done with Finland.

Also, I did need to add the two price books, "gbp-list-prices-sale" and
"gbp-list-prices"  to the RapalaEu site. I did not want to add this to
the data-impex as not to mess up the other instances.