RPS=430
New feature: When changing country from EU to UK, the currency changed
but not the flag icon. Now, when switching from EU to UK, the UK flag
will appear in the upper right. This will happen when changing the
country through the interstitial or the "change currency" link in the
upper right.