RPS-361
It seems, SFCC does not support the eval function - although it does
work on my sandbox. Because of this quirk, the interstitial window does
not pop up. 

I refactored how the URL are generated.Also had to edit the currency
selector and reedit the global preference 'ActiveCountries'.

These changes will also have to be merged into RPS-430 when this is
passed.