RPS-361
I've replaced the <isscript> that was creating the links. At first, I
did try to stick with this framework but in testing, it seemed that some
variables were not staying set. This new system seems to work way better
and is build for more future proof, easier to add more EU countries to
the mix. The new global preference, 'ActiveCountries', will hold a JSON
object that holds all the country info and that drives this
interstitial.