RPS-361
There is a new global preference, “Active Countries”, see below for
details.

I rewrote a lot of the code that creates and displays the interstitial
window. The original code was not really designed to allow one to switch
locales on products, contents, etc.and staying on that product, page,
etc. This new code will - but there are limits. Most of the content
asset pages should allow a user to switch pretty easily from EU to US
and back. The US and EU seem to share many of the same content asset
page names. Because the catalog between the US and EU are quite
different, a user switching from a category/product on the EU to US
might not get to the same page and will default back to homepage. Within
the EU, however, it’s different. As the catalog is the same across
sites, a user can easily switch from a category/product on the EU site
to the Finnish site and/or to the UK site and have the content
translated. And when switching between EU and UK the currency will
switch as well.

In the original code, the link to the Canadian site has only one URL so,
regardless what page a user is viewing in the EU or US, the link to
Canada will always go to the homepage. This issue was not addressed here
as it seemed way out of scope for this ticket.

A question I had when starting this code rework was, when Rapala adds
the UK to the site, where does that flag in order go? And anticipating a
French site, the same question arose. Currently, there is a global
preference to display the Finnish flag, “Show Finnish Flag” but this
seemed awkward, bulky when adding more and more countries. And add to
the fact, the order would have to be hard coded so I have created a new
global preference, “Active Countries”. This is a JSON object. Once all
the work is done to add a new locale to the site, to activate the flag
on the interstitial, edit this JSON object. With this approach, the
flags can also be sorted in the JSON - right in BM - no need to recode.
Once this code is approved, the global preference, “Show Finnish Flag”,
will no longer be needed.