RPS-377
I’ve fixed the button alignment for “Find Stores” by using the US site
as a base. There, the button is same height and alignment as the “Select
Radius” option. Miles are replaced for kilometers Also fixed an issue
with the site pref ( storeLookupUnit ) not being associated with an
attribute grouping. And I needed to import stores from CA from staging
for testing.