RPS-547
Weird event on staging. QA said the Est Tax was not showing on desktop
but showing on mobile. When I went to test on staging this morning, I
could see Est Tax on both desktop & mobile. I tested on my box and I
could not replicate the issue with same code. I did look on staging's
WebDav - current code version - and this file is there.Even the
cartridge path is correct, CA is to the left of EU and all the other
cartridges that use this template. So, my fix is, because this is Canada
only, I ripped out the Est Tax in the template.