RPS-361
There were still a few unforseen words that were not translated because
they were hard coded. A couple Finnish words, they wanted changed.Also,
on the No Hits page, they wanted the Best Sellers link removed because
it was not linking to anything. The link "failed" because, on the EU
site, there is no Best Sellers category. I've put a conditional around
it, if that category exists, display the link.