RPS-361
Added more Finish translations. A lot of these were not previously set
in reference files so they were hard coded right in either a template or
JavaScript file.

Removed the "Go To" text in the Button in the bottom of every content

There were still a few unforseen words that were not translated because
they were hard coded. A couple Finnish words, they wanted changed.Also,
on the No Hits page, they wanted the Best Sellers link removed because
it was not linking to anything. The link "failed" because, on the EU
site, there is no Best Sellers category. I've put a conditional around
it, if that category exists, display the link.

Fixed a few more hard coded translations.

Made text changes on the image zoom on pdp