RPS-361
Hopefully this is the last fix for the interstitial.

So, the reason why the flags sometimes get stacked vertically is, the
dialog window is so narrow and the scroll bar is different widths on
different browsers. I could not reduce the margins around the flag any
more because, if I did, there was not enough room to have the cart
symbol and the url both appear in one line under the respective country
flag.

I increased the width of the dialog box a bit, adjusted the width of the
flag containers and its left/right margin. 

I've tested this in Chrome/FF on Mac and also on IE 11.