RPS-525
I could not replicate any of the reopen issues. I'm suspecting, this
ticket was retested before the code was on staging. However, I did find
a small issue, the #footernew element did have a 'top: -100px' style
applied to it. Only on the homepage did this seem to have an effect and
on the other pages, this was overwritten by 'top: 0'. This style was
causing the footer to load under the bottom portion of the main
content.