RPS-430
On the PDP, fixed the issue of the borders and background colour around
the tabs section.

On the Interstitial, the URL displayed for the UK was
'rapala.com/en-GB'. It has not been switched to be 'rapala.co.uk'.

When switching to the UK, the currency will now switch to pounds. The
OnRequest was looking if there was a
request.httpPath.indexOf('SetLocale') or
request.httpPath.indexOf('Default'). If switching  from an EU site ( EU,
FI, etc ) to the UK, these values might not be there. I added to the
condition to see ig the request.locale is not default and if so, set the
new cookie and change the currency. 

I did find another bug. If a user was on the EU site and on the
interstitial, clicked on the US flag, the user would be redirected to
the US site no problem. However, if a user was on the UK or FI site, the
URL for the US would break. The URLUtils.url was setting the US primary
URL to the locale site ( for example http://rapala.fi ) and it was not
able to reset the language from the current locale to 'default'for the
US.