RPS-600 Desktop/Mobile - US/CA/EU - filters broken
fixed the issues with the filters. Switched "$('#main').on('load'," to
be "$('#main').load("

Also, in product-tile.js, on page load due to filters, the 'mouseleave'
function was firing off when it shouldn't and the script would try to
add data to the product thumb. this added to the console log errors -
not park of ticket but an easy fix