RPS-467
In the footer, I've replaced the US version of the Strikemaster logo
with the EU one. As the US is using a sprite ( one that includes the US
version of the Strikemaster logo ) and the EU is using stand alone brand
icons, there is no overlap here. 

All that needs to be done is the checkout logo issue - but i've put
notes in the ticket for that.it would require new sprite and some extra
coing - going above the 2 hour limit in this ticket