RPS-467
In the Recently Viewed carousel on the PDP, the Stikemaster logo wasn't
showing up. The code gets the logo name from the site name, in this case
Strikemaster is NOT camel-cased, however, the brand logo svg for
Strikemaster was camel-cased. Because of this, the names did not match
and the logo would not show. I've renamed the icon not to be
camel-cased.