RPS-611  Add new brand Marttiini to EU site
Added Martiini to the site. 

I had to create a new site pref, "activebrandmarttiini" in the attribute
group "ActiveBrands" so, if active, the brand will display the tab in
the main menu. Also, in the site pref, "brandtaborder", I had to add
'martiini|22" to the list of sites to be used on the EU site. 

On the homepage template, I had to add the
"brand-martinni-homepage-main" content slot.

And in SetCurrentSiteContext.ds, I had to add the new site, "martiini",
to the list of possible sites for the session. 

To be done - get a working catalog as the ones they sent ONLY have
products but no categories in the catalog have been created. Also, we
will need new logos. Both items have been mentioned in the ticket.