RPS-619 Update hreflang tags for FI and GB
This is set to ONLY affect the EU. Updated the hreflang tags for FI and
GB. In ViewHelpers.ds, I'm using the already defined global preference,
'ActiveCountries'. This has the sitePath, localID, and domain for each
of the active EU sites. If the site is active - that is added to this
global pref - the hreflang tag's URL will be written to point to the
correct EU site based on the language and set to the local language. If
a country is not listed in 'ActiveCountries', then there will be no
change. By using the global preference, this will be a dynamic change.