RPS-611
In the leftnavmenu template, the Marttiini site needed to be added to
the list of sites. 

I went through the rest of the site, found a few more places this might
happen. Fixed those. I wonder if getting rid of these repetitive <isif>
statements, making them into a check of a site pref holding a JSON
object might be better - fix it in one place afterwards. But, thats a
future project.