RPS-622
For "Product Item" - the SKU, in French, they wanted an adjustment,
removing the "No" ( number ) before the word "Article"

Also, the auto French locale seetings by SFCC were not correct and the
pricing format wasn't right. Set this up on dev and stage. put this fix
in bm-data so it is there for backup - import to other devs.